integration: extract test_named_bind_markers()
The test is extracted to `statements/named_bind_markers.rs`.
